docs: Update README File
Add a dataset that shows the customer list, transaction list, monthly total per customer, and 3 months summary it in.
Add instructions on how to invoke the API, how to run the unit tests, how to monitor the API, and how to run the dockerized API locally.

Related to Github Issue #1